Fix animation bug.
This limits animations' occurence to a specific breakpoint and adds
the appropriate class when the viewport is below that breakpoint.
This way, if a screen starts below the breakpoint and changes to
above the breakpoint the content will still appear (instead of being
hidden, like it was before this commit).